refactor: 디렉토리 정리
- dashboard 디렉토리 생성 후 이동
- TodoContent 상위 폴더로 이동
