+++ frontend/src/components/containers/dashboard/DashBoardContainer.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
-import Header from "../presentational/Header";
-import ToDoContent from "../presentational/dashboard/ToDoContent";
-import MainBox from "./MainBox";
+import Header from "../../presentational/Header";
+import ToDoContent from "../../presentational/ToDoContent";
+import MainBox from "../MainBox";
 
 export default function DashBoardContainer() {
   const mainTitle = "대시보드";
