+++ frontend/src/components/containers/metadata/MetaDataManagementContainer.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Header from "../../presentational/Header";
 import MainBox from "../MainBox";
-import ToDoContent from "../../presentational/dashboard/ToDoContent";
+import ToDoContent from "../../presentational/ToDoContent";
 
 export default function MetaDataManagementContainer(props) {
   const {mainTitle, tabNames} = props;
